requirements.txt
pandas==2.2.0
sqlmodel==0.0.14
mysql-connector-python==8.3.0
openpyxl==3.1.5
bcrypt==4.1.2
flet==0.24.1
python-dotenv==1.0.0
